requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pdfplumber
python-docx
langchain
langchain-openai
transformers
torch
streamlit
requests
